Add isZero method to Main and Refactor convertToInteger and add convert_IIV_To0Test
isZero checks if an integer is zero, this is useful since only invalid entries are zero

Refactor convertToInteger to account for repeated subtraction such as IIV or XXL, both of which are invalid but were not handled previously

Add convert_IIV_To0Test to test change to convertToInteger.  IIV previously returned 5, it now returns 0.